+++ packages/types/src/errors/NormalisableError.ts
@@ -1,0 +1,4 @@
+import { AxiosError } from 'axios';
+import { ApiError, ErrorWithMessage, PayloadError } from '.';
+
+export type NormalisableError = Error | AxiosError | PayloadError | ApiError | ErrorWithMessage | unknown;