+++ packages/types/src/errors/NormalisedError.ts
@@ -1,0 +1,9 @@
+export interface NormalisedError {
+  heading?: string;
+  message: string;
+  correlationId?: string;
+  status?: number;
+  code?: string;
+  validationErrors?: Record<string, string[]>;
+  disableMessage?: boolean;
+}