+++ packages/types/src/errors/ValidationError.ts
@@ -1,0 +1,9 @@
+import { ApiErrorPayload } from '.';
+
+export interface ValidationErrorPayload extends ApiErrorPayload {
+  errors: Record<string, string | string[]>;
+}
+
+export interface ValidationError {
+  payload: ValidationErrorPayload;
+}