+++ apps/docs/CHANGELOG.md
@@ -1,1 +1,16 @@
 # @batoanng/docs
+
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @batoanng/utils@2.1.0
+  - @batoanng/core@1.0.0
+
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @batoanng/utils@2.0.0