+++ packages/components/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @batoanng/components
 
+## 2.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @batoanng/tsconfig@1.2.0
+  - @batoanng/utils@2.1.0
+
 ## 2.0.1
 
 ### Patch Changes