+++ packages/tsconfig/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @batoanng/tsconfig
 
+## 1.2.0
+
+### Minor Changes
+
+- pump lib version
+
 ## 1.1.1
 
 ### Patch Changes