+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @batoanng/utils
 
+## 2.1.0
+
+### Minor Changes
+
+- pump lib version
+
+## 2.0.0
+
+### Major Changes
+
+- Update deps and build package
+
 ## 1.0.0
 
 ### Major Changes